requirements.txt
Flask
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib